refactor: family module package cleanup
